+++ src/loaders/index.ts
@@ -1,5 +1,10 @@
 export { Asset } from "./Asset.js";
 export { AssetManager } from "./AssetManager.js";
-export { JSONSceneLoader, type SerializedObject, type SerializedScene } from "./JSONSceneLoader.js";
+export {
+  JSONSceneLoader,
+  type SerializedBufferGeometry,
+  type SerializedObject,
+  type SerializedScene,
+} from "./JSONSceneLoader.js";
 export { Loader } from "./Loader.js";
 export { TextureLoader } from "./TextureLoader.js";